Safari-609.1.11.1/LayoutTests/fast/dom/getElementsByClassName/case-sensitive-expected.txt

PASS getByClassName('a') is "a"
PASS getByClassName('A') is "A"
PASS getByClassName('ä') is "ä"
PASS getByClassName('Ä') is "Ä"
PASS successfullyParsed is true

TEST COMPLETE
